Fix issues with Op Force

+++ Projects/Android/jni/src/Lambda1VR/VrCommon.h
@@ -37,7 +37,7 @@ extern float playerHeight;
 extern float playerYaw;
 
 extern bool showingScreenLayer;
-extern float vrFOV;
+extern int vrFOV;
 
 extern vec3_t worldPosition;
